+++ lib/flutter_refresh_indicator.dart
@@ -1,7 +1,5 @@
 library;
 
-export 'components/refresh_indicator_style.dart';
-
 export 'widgets/clamping_refresh_indicator.dart';
 export 'widgets/bouncing_refresh_indicator.dart';
 export 'widgets/refresh_indicator.dart';