+++ lib/widgets/primary_refresh_indicator.dart
@@ -1,25 +1,26 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_refresh_indicator/components/refresh_indicator_style.dart';
+import 'package:flutter_refresh_indicator/flutter_refresh_indicator.dart';
 
 /// A inherited widget globally provides the style information for
 /// refresh indicators to all descendant widgets in the widget tree.
 class PrimaryRefreshIndicator extends InheritedWidget {
   const PrimaryRefreshIndicator({
     super.key,
-    this.clamping,
-    this.bouncing,
+    this.clampingIndicatorBuilder,
+    this.bouncingIndicatorBuilder,
     required super.child,
   });
 
-  /// The style that will be defined to refresh indicator for [ClampingScrollPhysics].
-  final RefreshIndicatorStyle? clamping;
+  /// The builder that will be defined to refresh indicator for [ClampingScrollPhysics].
+  final ClampingRefreshIndicatorBuilder? clampingIndicatorBuilder;
 
-  /// The style that will be defined to refresh indicator for [BouncingScrollPhysics].
-  final RefreshIndicatorStyle? bouncing;
+  /// The builder that will be defined to refresh indicator for [BouncingScrollPhysics].
+  final BouncingRefreshIndicatorBuilder? bouncingIndicatorBuilder;
 
   @override
   bool updateShouldNotify(PrimaryRefreshIndicator oldWidget) {
-    return oldWidget.clamping != clamping || oldWidget.bouncing != bouncing;
+    return oldWidget.clampingIndicatorBuilder != clampingIndicatorBuilder ||
+        oldWidget.bouncingIndicatorBuilder != bouncingIndicatorBuilder;
   }
 
   /// Finds the [PrimaryRefreshIndicator] from the closest instanc